Move chatbot tools to chatbot module level
Used `git mv` to move `memory.py`, `skills.py`, and `todo.py` from `plugin/modules/chatbot/tools/` directly into `plugin/modules/chatbot/`. Removed the empty `tools` directory and its `__init__.py`.

Updated references in `document.py` and `test_agent_memory_skills.py` to point to the new paths. Also wrapped LibreOffice-specific `com.sun.star` exceptions in a `try/except ImportError` block in `tool_loop.py` to ensure headless non-uno test environments continue to function properly. Included `"com"` in expected mock failure messages in `test_i18n.py`.
