+++ plugin/modules/chatbot/memory.py
