+++ plugin/modules/chatbot/skills.py
