--- plugin/modules/chatbot/tools/__init__.py
@@ -1,6 +1,0 @@
-# Placeholder tools package for chatbot-specific ToolBase implementations.
-#
-# NOTE: The experimental planning/todo tool is currently commented out in
-# todo.py so that it is not discovered or exposed to LLMs. See that file
-# for details on how to enable it.
-
